# [Feature] Speculative decoding support lookahead