--- a/python/sglang/srt/managers/schedule_batch.py
+++ b/python/sglang/srt/managers/schedule_batch.py
@@ -836,7 +836,6 @@ class ScheduleBatch(ScheduleBatchDisaggregationDecodeMixin):
     token_to_kv_pool_allocator: BaseTokenToKVPoolAllocator = None
     tree_cache: BasePrefixCache = None
     is_hybrid: bool = False
-    eos_token_id: int | None = None
 
     # Batch configs
     model_config: ModelConfig = None
@@ -971,13 +970,6 @@ class ScheduleBatch(ScheduleBatchDisaggregationDecodeMixin):
             is_prefill_only=all(req.is_prefill_only for req in reqs),
             chunked_req=chunked_req,
         )
-
-    @property
-    def eos_id(self) -> int:
-        # TODO: support multiple eos token ids
-        if self.eos_token_id is None and len(self.reqs) > 0:
-            self.eos_token_id = self.reqs[0].tokenizer.eos_token_id
-        return self.eos_token_id
 
     def batch_size(self):
         return len(self.reqs)
@@ -1795,7 +1787,6 @@ class ScheduleBatch(ScheduleBatchDisaggregationDecodeMixin):
             token_type_ids=self.token_type_ids,
             spec_algorithm=self.spec_algorithm,
             spec_info=self.spec_info,
-            eos_id=self.eos_id,
             hicache_consumer_index=self.hicache_consumer_index,
             capture_hidden_mode=(
                 CaptureHiddenMode.FULL
@@ -1885,7 +1876,6 @@ class ScheduleBatch(ScheduleBatchDisaggregationDecodeMixin):
 class ModelWorkerBatch:
     # The batch id
     bid: int
-    eos_id: int
     # The forward mode
     forward_mode: ForwardMode
     # The input ids