--- a/python/sglang/srt/speculative/cpp_lookahead/.clang-format
+++ b/python/sglang/srt/speculative/cpp_lookahead/.clang-format
@@ -1,0 +1,1 @@
+../../../../../sgl-kernel/.clang-format