… python/pyproject.toml
… python/sglang/srt/model_executor/cuda_graph_runner.py
… python/sglang/srt/model_executor/model_runner.py
… python/sglang/srt/server_args.py
… python/sglang/srt/speculative/cpp_lookahead/lookahead.cpp
… python/sglang/srt/speculative/cpp_lookahead/lookahead.h
… python/sglang/srt/speculative/cpp_lookahead/lookahead_cache.py
… python/sglang/srt/speculative/cpp_lookahead/lookahead_cache_binding.cpp
… python/sglang/srt/speculative/cpp_lookahead/param.h
… python/sglang/srt/speculative/cpp_lookahead/queue.h
… python/sglang/srt/speculative/eagle_utils.py
… python/sglang/srt/speculative/eagle_worker.py
… python/sglang/srt/speculative/lookahead_utils.py
… python/sglang/srt/speculative/lookahead_worker.py
… python/sglang/srt/speculative/spec_info.py
… python/sglang/test/test_utils.py
… sgl-kernel/CMakeLists.txt
… sgl-kernel/csrc/common_extension.cc
… sgl-kernel/csrc/common_extension_rocm.cc
… sgl-kernel/csrc/speculative/lookahead_utils.cu
… sgl-kernel/include/sgl_kernel_ops.h
… sgl-kernel/python/sgl_kernel/__init__.py
… sgl-kernel/python/sgl_kernel/speculative.py
… sgl-kernel/setup_rocm.py
… sgl-kernel/tests/speculative/test_lookahead_utils.py
… test/srt/run_suite.py
… test/srt/test_lookahead_speculative_decoding.py