<!-- Thank you for your contribution! Please follow these guidelines to enhance your pull request. If anything is unclear, submit your PR and reach out to maintainers for assistance. Join our Slack community at https://slack.sglang.ai to discuss further. -->

## Motivation

<!-- Describe the purpose and goals of this pull request. -->
For large language model with good output locality, using lookahead can achieve significant performance improvement at a relatively low cost. This PR references [#2790 ](https://github.com/sgl-project/sglang/pull/2790).
## Modifications

<!-- Detail the changes made in this pull request. -->

## Accuracy Tests

<!-- If this pull request affects model outputs (e.g., changes to the kernel or model forward code), provide accuracy test results. -->
This feature has been running in multiple of our [business scenarios](https://www.wenxiaobai.com/) for over two months, and its stability and acceleration effect have been verified.

## Benchmarking and Profiling

<!-- If this pull request impacts inference speed, provide benchmarking and profiling results. -->
### The test results on our own model
 We apply this feature to smaller-scale language models with parameter sizes, for use in functions like user intent recognition and behavior routing. These models are characterized by short output and high locality, making them quite suitable for lookahead. The image below shows the accepted length of each forward pass after using lookahead.

<img width="3826" height="1776" alt="image" src="https://github.com/user-attachments/assets/42558c94-ed10-458c-8576-6e52a4c63899" />

The optimal acceleration effect is as follows, with an average accepted length of approximately 4.8 and num prompt 256:

| concurrency | base       | lookahead | speedup |
|--------------|---------|------------|----------|
| 1                     | 413.14   |     1079.52   |     2.61     |
| 2                     | 672.43  |     1489.14   |     2.21    |
| 3                     | 859.09  |     150.29    |     1.92    |
| 4                     | 996.53  |     1799.86  |    1.81     |

### Acceleration effect on the public test set
model: [Qwen2.5-Coder-7B-Instruct](https://huggingface.co/Qwen/Qwen2.5-Coder-7B-Instruct)
dataset: [qwen2.5_test_python](https://huggingface.co/datasets/ZAAAAAC/qwen2.5_test_python)
num prompt: 1024
NOTE: **Before the test, need to convert the format from "parque" to "sharedgpt".**

```bash
# h20, tp 2
python3 -m sglang.launch_server \
    --host ${h20_4} \
    --port 33337 \
    --model-path ${Model_Path} \
    --tp 2 \
    --mem-fraction-static 0.9 \
    --max-prefill-tokens 131072 \
    --log-level info \
    --speculative-algorithm ${speculative_algo}  --speculative-num-draft-tokens 16 \
    --speculative-lookahead-min-match-window-size 1 \
    --speculative-lookahead-max-match-window-size 16 \
    --speculative-lookahead-min-bfs-breadth 1 \
    --speculative-lookahead-max-bfs-breadth 10 \
    --speculative-lookahead-branch-length 18 \
    --trust-remote-code 2>&1 | tee ${0}.log
```

```bash
# bench
declare -a request_rates=(256 256 256 256)
declare -a max_concurrency=(1 2 3 4)
declare -a num_prompts=(1024 1024 1024 1024)

for i in "${!request_rates[@]}"; do
    RATE=${request_rates[$i]}
    CONCURRENCY=${max_concurrency[$i]}
    PROMPTS=${num_prompts[$i]}

    echo "Running benchmark with request rate $RATE, concurrency $CONCURRENCY, prompts $PROMPTS"
    python3  /sgl-workspace/sglang/python/sglang/bench_serving.yuanshi.py \
        --apply-chat-template \
        --dataset-name sharegpt \
        --dataset-path ${dataset_path} \
        --disable-ignore-eos \
        --backend sglang \
        --flush-cache \
        --model $Model_Path \
        --sharegpt-output-len 9999 \
        --request-rate $RATE \
        --max-concurrency $CONCURRENCY \
        --num-prompts $PROMPTS \
        --host $addr --port $port
    wait
done
```

| concurrency | base | lookahead | accept | speedup |
|--------------|------|-----------|--------|----------|
| 1                     | 265   |     429       |     1.45 |      1.62    |
| 2                     | 507   |     782       |     1.47 |      1.54    |
| 3                     | 746   |     987       |     1.43 |      1.32    |
| 4                     | 971   |     1126      |    1.43  |      1.16     |

## Question
Compared with the full-type tree mask, the qlen-type mask offers an 8% performance improvement. However, the qlen-type mask requires support from FlashInfer. Is there any plan for FlashInfer to support this feature?

## Checklist

- [x] Format your code according to the [Format code with pre-commit](https://docs.sglang.ai/developer_guide/contribution_guide.html#format-code-with-pre-commit).
- [x] Add unit tests according to the [Run and add unit tests](https://docs.sglang.ai/developer_guide/contribution_guide.html#run-and-add-unit-tests).
- [ ] Update documentation according to [Write documentations](https://docs.sglang.ai/developer_guide/contribution_guide.html#write-documentations).
- [ ] Provide accuracy and speed benchmark results according to [Test the accuracy](https://docs.sglang.ai/developer_guide/contribution_guide.html#test-the-accuracy) and [Benchmark the speed](https://docs.sglang.ai/developer_guide/contribution_guide.html#benchmark-the-speed).