@Swipe4057 2025-09-05
does performance drop with large batch sizes?