@Swipe4057 2025-09-05
> > does performance drop with large batch sizes?
> 
> Yes, there will be a loss in performance. The optimal implementation is to reduce the number of draft tokens as the batch size increases.

Is there compatibility with speculative_token_map parameter? It improved latency on large batch sizes for eagle mode