@a4zhangfei 2025-09-05
@ispobock hi~ I have refactored the verify logic to maximize the reuse of the eagle verify logic. Please review it when you have time. Thanks