@a4zhangfei 2025-09-09
> Awesome work! Is the lookahead feature in your implementation based on this paper: [Break the Sequential Dependency of LLM Inference Using Lookahead Decoding](https://arxiv.org/abs/2402.02057)?

@valorix25 Thanks! Yes, this implementation is based on this paper.