@valorix25 2025-09-09
Hi there, After pulling your latest code updates today, I’ve been consistently encountering CUDA errors. However, the code was running perfectly fine for me before. Do I need to rebuild the sgl-kernel? If so, could you explain why this step is now necessary?

Below is the error message：
```
[2025-09-09 11:53:36] Prefill batch. #new-seq: 1, #new-token: 3555, #cached-token: 0, token usage: 0.00, #running-req: 0, #queue-req: 0, 
[2025-09-09 11:53:37] Prefill batch. #new-seq: 3, #new-token: 58, #cached-token: 10604, token usage: 0.43, #running-req: 1, #queue-req: 4, 
[2025-09-09 11:53:37] Scheduler hit an exception: Traceback (most recent call last):
  File "sglang/python/sglang/srt/managers/scheduler.py", line 2622, in run_scheduler_process
    scheduler.event_loop_normal()
  File "miniconda3/envs/sglang/lib/python3.10/site-packages/torch/utils/_contextlib.py", line 120, in decorate_context
    return func(*args, **kwargs)
  File "sglang/python/sglang/srt/managers/scheduler.py", line 830, in event_loop_normal
    result = self.run_batch(batch)
  File "sglang/python/sglang/srt/managers/scheduler.py", line 1842, in run_batch
    ) = self.draft_worker.forward_batch_speculative_generation(batch)
  File "sglang/python/sglang/srt/speculative/lookahead_worker.py", line 227, in forward_batch_speculative_generation
    logits_output, next_token_ids, num_accepted_tokens = verify_input.verify(
  File "sglang/python/sglang/srt/speculative/lookahead_utils.py", line 397, in verify
    self._greedy_verify(batch, logits_output)
  File "sglang/python/sglang/srt/speculative/lookahead_utils.py", line 249, in _greedy_verify
    target_predict = torch.argmax(logits_output.next_token_logits, dim=-1)
torch.AcceleratorError: CUDA error: an illegal memory access was encountered
Compile with `TORCH_USE_CUDA_DSA` to enable device-side assertions.


[2025-09-09 11:53:37] Received sigquit from a child process. It usually means the child failed.
```