@a4zhangfei 2025-09-14
> Thanks!
> 
> 1. May you add an e2e test for lookahead speculative decoding, like [test_standalone_speculative_decoding.py](https://github.com/sgl-project/sglang/blob/main/test/srt/test_standalone_speculative_decoding.py).
> 2. Have you verified this on other attention backends, or only flashinfer?
> 3. Is this compatible with other features, (e.g., dp attention)?

@Qiaolin-Yu Hi~, I have added the e2e test for [lookahead speculative decoding](https://github.com/a4zhangfei/sglang/blob/lookahead/test/srt/test_lookahead_speculative_decoding.py).