@a4zhangfei 2025-09-14
> May you add the test to `test/srt/run_suite.py` to enable it in CI?

ok, it's done! CI added.