@Qiaolin-Yu 2025-09-15
For reference:
Results for llama3.3-70b, bs1, mt-bench
- without spec: 63.31
- lookahead (default params): 79.34, acc len: 1.66
- standalone (num_steps=8, topk=8 draft_num_token=64): 180.43, acc len: 6.21
- eagle3 (num_steps=8, topk=8 draft_num_token=64): 209.55, acc len: 5.55