@a4zhangfei 2025-09-16
> ```shell
> # h20, tp 2
> python3 -m sglang.launch_server \
>     --host ${h20_4} \
>     --port 33337 \
>     --model-path ${Model_Path} \
>     --tp 2 \
>     --mem-fraction-static 0.9 \
>     --max-prefill-tokens 131072 \
>     --log-level info \
>     --speculative-algorithm ${speculative_algo}  --speculative-num-draft-tokens 16 \
>     --speculative-lookahead-min-match-window-size 1 \
>     --speculative-lookahead-max-match-window-size 16 \
>     --speculative-lookahead-min-bfs-breadth 1 \
>     --speculative-lookahead-max-bfs-breadth 10 \
>     --speculative-lookahead-branch-length 18 \
>     --trust-remote-code 2>&1 | tee ${0}.log
> ```

> @a4zhangfei Could you provide the exact commands to reproduce the qwen-coder performance test results?

More test data is needed so that the acceptance length of speculative sampling can be improved. For example, 2048 prompts.