@Qiaolin-Yu 2025-09-16
> @Qiaolin-Yu By the way, the acceptable length I provided is calculated on the caller side based on the number of returned tokens, which is more accurate. The statistics from SGLang are not accurate enough and tend to be slightly higher.

Could you check slack message 👀, need more offline discussion.