@a4zhangfei 2025-09-16
@reyoung I referred to this [code](https://github.com/sgl-project/sglang/blob/main/python/sglang/srt/mem_cache/cpp_radix_tree/radix_tree.py#L10): 

```
lookahead_cache_cpp = load(
    name="lookahead_cache_cpp",
    sources=[
        f"{_abs_path}/lookahead_cache_binding.cpp",
        f"{_abs_path}/lookahead.cpp",
    ],
    extra_cflags=["-O3", "-std=c++20"],
)
```