@a4zhangfei 2025-09-17
> > I referred to this [code](https://github.com/sgl-project/sglang/blob/main/python/sglang/srt/mem_cache/cpp_radix_tree/radix_tree.py#L10):
> 
> However, the cpp/h files are missing when using `pip install` and in the wheel package.

Ok~, thanks, I will fix it